Add "safe_name" to context and templates.
This allows the widget to be used with an admin inline where the name has hyphens that get interpreted as an operator in Javascript.  This changeset also updates methods to be specific to the single widget being outputted since there can be multiple on the page as an inline.  (Still a couple bugs here to be fleshed out.)