Add optional parameter (--dir) to sync_media_s3.
The --dir/-d parameter allows the end user to specify a different path other than MEDIA_ROOT with which to push to S3.  My use case for this was using django-compress to compile down some of my media, which I wanted to host on S3, but didn't want to move the rest of my media (images and flash files).