Make ForeignKeySearchInput usable outside changeforms
Previously the `related_url` variable was calculated assuming the page with a widget is nested 3 levels down (for example, a model's changeform). That was making it harder to use the widget in a custom admin views, which could reside at the arbitrary nesting level. Here we switch to using a more flexible `reverse()` call.