Fix admin_generator incompatible "default" identifier raising TypeError.
In admin_generator.py we change to the argparse "default" string replace
style and update ProxyParser to handle the necessary replacement for
optparse. The 3.5.1 docs for argparse state: "The help strings can
include various format specifiers to avoid repetition of things like the
program name or the argument default. The available specifiers include
the program name, %(prog)s and most keyword arguments to add_argument(),
e.g. %(default)s, %(type)s, etc." The use of "most" is ambiguous and in
preliminary testing, this functionality is at least missing for the
kwarg "action". It is not the intent of CompatibilityBaseCommand or
ProxyParser to add functionality to optparse, so we make no effort to
handle whatever other kwargs may be supported as formatted
identifiers in argparse, as the optparse docs only mention %default
in the help string. %prog exists in optparse only in the usage and
version strings.