Fix settings.GRAPH_MODELS breaking graph_models.
In 160aaa8, graph_models.Command.options_from_settings was left out of
the change to introduce CompatibilityBaseCommand. To resolve this, move
our arguments into an instance variable that we can use to override
defaults in __init__() with settings.GRAPH_MODELS before adding
arguments. This removes any optparse specific code and pre-empts the
need for any argpase-specific processing after add_arguments(). Remove
optparse NO_DEFAULT import as the usage is not clearly documented by
optparse and would seemingly lack consequence in any scenario short of
someone editing the file for optparse-specific functionality (which
should not be happening).