Update graph_model css to match 2018 style djangoproject.com
First cut of an attempt to update the default style from the ~2005 version of djangoproject to the current style.

Removes pale golden rod and olive green for djangoproject green and white. Also uses Roboto, which is a Google Font, but for some output formats would be required to be installed natively on the system (as would Helvetica Bold, which is not on maOS Mojave, causing some font issues when generating SVGs)