Fix show_urls --decorator
`decorator` initially contains the list of decorators that show_urls
should look for. However, it was overwritten with the the formatted
decorators further down the loop body. As a consequence, starting from
the second loop iteration no decorators were found anymore.

Now, the variable `decorators` holds the formatted decorator information.