Add a py.typed file to export type annotations
In order for any of the type annotations used within django-extensions to be
available to downstream users, PEP-561 requires that an empty "py.typed"
marker file be present as package data. For more details, see:
https://blog.whtsky.me/tech/2021/dont-forget-py.typed-for-your-typed-python-package/

This allows downstream users to have type checking when directly using any
Python APIs which happen to have type annotations.

More importantly, it allows Django and DRF mixin classes (like
"ModelUserFieldPermissionMixin") and subclasses (like "TimeStampedModel") to
inherit type information if type stub packages for Django or DRF are in use.